Test for proposal verification
